requirements.txt
python-jose~=3.5.0
fastapi~=0.115.13
python-dotenv~=0.21.0
pytesseract~=0.3.13
pillow~=10.4.0
exifread~=3.3.1
better-profanity~=0.7.0
weasyprint~=65.1